Sync issue

+++ Pdf_In_Browser_1/Default.aspx.cs
@@ -36,12 +36,14 @@ namespace Pdf_In_Browser_1
             //Deletes the contents of the TestImages Folder when LoadPDF button is clicked
             System.IO.DirectoryInfo directory = new DirectoryInfo(Server.MapPath("~/TestImages"));
 
-            foreach (FileInfo file in directory.GetFiles())
+            if(directory.GetFiles().Length > 0)
             {
-                file.Delete();
+                foreach (FileInfo file in directory.GetFiles())
+                {
+                    file.Delete();
+                }
             }
-
-
+            
             if (FileUpload1.HasFile)
             {
                 //Saves the document, updates page label, and then fires client side function to make API call to save all pages as images
@@ -51,9 +53,6 @@ namespace Pdf_In_Browser_1
                 UpdatePageTotalUI(controller.GetPageCount());
                 Page.RegisterStartupScript("page", "<script language='javascript'>saveFirstPages()</script>");
             }
-            
-
-
         }
     }
 }
